requirements.txt
django==1.4.3
django-nose
pep8
psycopg2==2.4.5
pyflakes
python-dateutil
git+https://github.com/wesokes/django-query-builder.git@0.3.2
pytz==2012h
